requirements.txt
python-telegram-bot>=12.4
plotly==4.9.0